Move `StoryCachePolicy` to its own file
