+++ HackrNews/Hackr News Feed Cache/Story/StoryCachePolicy.swift
@@ -1,0 +1,15 @@
+//
+//  Copyright © 2021 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//
+
+import Foundation
+
+struct StoryCachePolicy {
+    private static let calendar = Calendar(identifier: .gregorian)
+    private static var maxCacheAgeInMinutes: Int { 5 }
+
+    static func validate(_ timestamp: Date, against date: Date) -> Bool {
+        guard let maxCacheAge = calendar.date(byAdding: .minute, value: maxCacheAgeInMinutes, to: timestamp) else { return false }
+        return date < maxCacheAge
+    }
+}
